pyproject.toml
[project]
name = "neo4j-text2cypher-agent"
version = "0.1.0"
description = "Modern Text2Cypher agent using LangGraph 1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grandalf>=0.8",
    "langchain-core>=0.3.76",
    "langchain-neo4j>=0.5.0",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.8",
    "neo4j>=5.28.2",
    "nicegui>=3.0.0",
    "pandas>=2.2.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
